refactor common test case logic
